+++ packages/vendure-plugin-sendcloud/src/index.ts
@@ -3,3 +3,4 @@ export * from './api/sendcloud.handler';
 export * from './api/additional-parcel-input-items';
 export * from './api/types/sendcloud.types';
 export * from './api/types/sendcloud-api.types';
+export * from './config/fulfill-settled-orders-task';