+++ packages/vendure-plugin-qls-fulfillment/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.7.0 (2026-03-15)
+
+- Pass order phone number to QLS
+
+# 1.6.0 (2026-03-09)
+
+- Skip syncing disabled products/variants to QLS
+
+# 1.5.1 (2026-02-18)
+
+- Pass error message to `QlsVariantSyncFailedEvent` instead of the error object.
+
 # 1.5.0 (2026-02-18)
 
 - Allow passing additional order items with `addAdditionalOrderItems()` that will be added to the QLS order.